Refactor AsyncPromise catch method and enhance awaitPromise function for better promise handling
