+++ tests/Pest.php
@@ -2,6 +2,9 @@
 
 declare(strict_types=1);
 
+use React\EventLoop\Loop;
+use React\Promise\PromiseInterface;
+
 /*
 |--------------------------------------------------------------------------
 | Test Case
@@ -41,7 +44,44 @@ expect()->extend('toBeOne', function () {
 |
 */
 
-function something()
+/**
+ * Runs the event loop until the given promise settles or a timeout is reached.
+ *
+ * @param  PromiseInterface<mixed, \Throwable>  $promise  The promise to await.
+ * @param  float  $timeout  Maximum time in seconds to wait.
+ * @return mixed The resolved value of the promise.
+ *
+ * @throws \Throwable If the promise rejects.
+ */
+function awaitPromise(PromiseInterface $promise, float $timeout = 2.0)
 {
-    // ..
+    $resolved = false;
+    $rejected = false;
+    $result = null;
+    $error = null;
+
+    $promise->then(
+        function ($val) use (&$resolved, &$result) {
+            $resolved = true;
+            $result = $val;
+            Loop::stop();
+        },
+        function ($err) use (&$rejected, &$error) {
+            $rejected = true;
+            $error = $err;
+            Loop::stop();
+        }
+    );
+
+    Loop::addTimer($timeout, function () {
+        Loop::stop();
+    });
+
+    Loop::run();
+
+    if ($rejected) {
+        throw $error;
+    }
+
+    return $result;
 }
